Added two-way experimental modes (Not currently working), added security functionality for gamepads, etc.

+++ IOSync/src/devices/keyboard.h
@@ -42,7 +42,7 @@ namespace iosync
 		#ifdef PLATFORM_WINDOWS
 			typedef DWORD keyboardKey;
 		#else
-			typedef unsigned long keyboardKey;
+			typedef unsigned int keyboardKey;
 		#endif
 
 		typedef deque<keyboardAction> keyboardActionQueue;
